correção inserção alunoatividade

+++ src/main/java/br/com/unincor/webSite/model/domain/Aluno.java
@@ -8,7 +8,6 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Getter
 @Setter
@@ -20,7 +19,7 @@ import lombok.ToString;
 public class Aluno implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
     private String email;
     private String senha;
